refactor: extract tracing types into tracing/types.ts
