+++ src/types.ts
@@ -7,6 +7,7 @@ import type {
   LanguageModelUsage,
 } from "ai";
 import type { GuardrailResult } from "@/guardrail/types";
+import type { TracingConfig } from "@/tracing/types";
 
 // ---------------------------------------------------------------------------
 // Context
@@ -30,36 +31,6 @@ export interface RunContext<TContext = unknown> {
  * maxOutputTokens, topP, seed, retries, timeout, etc. at the agent level.
  */
 export type ModelSettings = CallSettings;
-
-// ---------------------------------------------------------------------------
-// Tracing
-// ---------------------------------------------------------------------------
-
-export interface TraceSpan {
-  traceId: string;
-  spanId: string;
-  parentSpanId?: string;
-  name: string;
-  type: "agent" | "llm" | "tool" | "guardrail" | "handoff" | "custom";
-  startTime: number;
-  endTime?: number;
-  attributes: Record<string, unknown>;
-}
-
-export interface TraceProcessor {
-  onTraceStart(traceId: string): void | Promise<void>;
-  onSpan(span: TraceSpan): void | Promise<void>;
-  onTraceEnd(traceId: string): void | Promise<void>;
-}
-
-export interface TracingConfig {
-  enabled?: boolean;
-  traceId?: string;
-  groupId?: string;
-  processors?: TraceProcessor[];
-  redactInput?: boolean;
-  redactOutput?: boolean;
-}
 
 // ---------------------------------------------------------------------------
 // Run hooks
